test: better naming for test library files
