--- src/shared/test-email.ts
@@ -1,57 +1,0 @@
-import fetch, { Response } from 'node-fetch'
-
-import { SMTP_HOST } from '@shared/config'
-
-interface MailhogEmailAddress {
-  Relays: string | null
-  Mailbox: string
-  Domain: string
-  Params: string
-}
-interface MailhogMessage {
-  ID: string
-  From: MailhogEmailAddress
-  To: MailhogEmailAddress[]
-  Content: {
-    Headers: {
-      'Content-Type': string[]
-      Date: string[]
-      From: string[]
-      'MIME-Version': string[]
-      'Message-ID': string[]
-      Received: string[]
-      'Return-Path': string[]
-      Subject: string[]
-      To: string[]
-      [key: string]: string[]
-    }
-    Body: string
-    Size: number
-  }
-  Created: string
-  Raw: {
-    From: string
-    To: string[]
-    Data: string
-  }
-}
-
-export interface MailhogSearchResult {
-  total: number
-  count: number
-  start: number
-  items: MailhogMessage[]
-}
-
-export const mailHogSearch = async (query: string, fields = 'to'): Promise<MailhogMessage[]> => {
-  const response = await fetch(
-    `http://${SMTP_HOST}:8025/api/v2/search?kind=${fields}&query=${query}`
-  )
-  return ((await response.json()) as MailhogSearchResult).items
-}
-
-export const deleteMailHogEmail = async ({ ID }: MailhogMessage): Promise<Response> =>
-  await fetch(`http://${SMTP_HOST}:8025/api/v1/messages/${ID}`, { method: 'DELETE' })
-
-export const deleteEmailsOfAccount = async (email: string): Promise<void> =>
-  (await mailHogSearch(email)).forEach(async (message) => await deleteMailHogEmail(message))
